mmc: sdhci-of-arasan: add support for EMMC power control
This commit introduces EMMC power control for XSC Q8 RevA.

The sdhci device exposes two new attributes to the sysfs:
* pwr_en (1:power on, 0:power off)
* state (on, off, oc_1v8:overcurrent 1V8, oc_3v3:overcurrent 3V3)

The pwr_en attribute can be used to enable/disable the EMMC.
When an overcurrent is detected, the EMMC is disabled.

Disabling the EMMC suspends the SDHCI host, set the power control GPIO,
deactivate the clocks and forces the EMMC pins to tristate.

An overcurrent interrupt read the overcurrent state before disabling the
EMMC, but does not suspend the SDHCI as this operation is not allowed in
the atomic context.