spi-nor: Do not modify the supported devices list
The list defines all basic information about the NOR devices. During the
scan sequence, some parameters are updated.

This behavior make the driver crash when it is reloaded. With this
patch, the driver doesn't modify the const list of device but instead,
saves the modified parameters in volatile structures that are
reallocated during the scan sequence.

Signed-off-by: Detlev Casanova <dec@xiphos.com>