mmc: sdhci-of-arasan: pull down MIO lines when powered down
The eMMC MIO lines shall be pulled down when the eMMC is powered off.
This commit adds a pull_down_timer to pull_down the lines 500ms after
that the MMC is powered off.

This timer has been set to enable the mmc core to detect the removal
of the card and release the SDHCI host. This avoid the following error:

  irq 32: nobody cared [...]

  handlers:
  [<ffffff80086288e8>] sdhci_irq threaded [<ffffff8008627368>] sdhci_thread_irq
  Disabling IRQ #32

When the pull down registers are set too soon, an IRQ is raised, is
ignored by the sdhci host driver, and the eMMC can not be powered on
again.