iio: jesd204: xilinx_transceiver: Fix CPLL VCO range
This patch fixes the VCO ranges used for CPLL transceivers. The VCO range
varies with voltage only for GTH type transceivers, while both GTH and GTY
are limited by the speed grade.

For speed grade 1 the VCO is limited to 4.25 GHz.
If the tranceivers are GTH and the voltage is below 0.85 V then VCO is
limited to 4.25 GHz.

Fixes: 4b66868b9505 (iio: jesd204: axi_xcvr: Add support for fpga info)
Signed-off-by: Mircea Caprioru <mircea.caprioru@analog.com>