iio: jesd204: xilinx_transceiver: simplify QPLL math during comparison
This is a minor optimization, but a slightly bigger optimization when
reading the code. Whenever I am re-reading this code, it takes me a while
to re-understand.

Ref: https://www.xilinx.com/support/documentation/user_guides/ug476_7Series_Transceivers.pdf
Page: 55

Vco_Freq is computed as:
  Vco_Freq = (refclk_khz * n) / m
Linerate computed as:
  LineRate = Vco_Freq / d

Make sure to not confuse Vco_Freq with fPLLClkout.
   fPLLClkout = (refclk_khz * n) / (m * 2), so technically
   Vco_Freq = 2 * fPLLClkout
And the 2 is reduced in both equations.

For GTY4:
Ref: https://www.xilinx.com/support/documentation/user_guides/ug578-ultrascale-gty-transceivers.pdf
Page: 49
For GTY4: LineRate = (2 * Vco_Freq) / d

For QPLL, try to match lane-rate at half, then at full-rate for GTY4
transceivers.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>