spi-nor: Add support for ISSI chips
ISSI NOR chips do not have the BP3 bit at the same place in the Status
Register. This adapts the code to support both possible configurations.

Also, the TB bit is in the Function register, which is not supported
currently. (Bottom configuration is disabled)
This commit makes sure that SR_TB is not modified when HAS_SR_TB is
false (which would actually modify the value of BP3).

Signed-off-by: Detlev Casanova <dec@xiphos.com>