xilinx_dma: Use the number of SG segments in coalesce
The Xilinx implementation uses the number of pending DMA descriptors but
doesn't take into account the scatter gather segments.

What happens in this situation is that there is an interrupt after each
SG transfer. A driver could count the interrupts but sometimes, they
come too fast and some of them are missed.

To prevent that, the IRQThreshold field is used to get only 1 interrupt
after all SG segments have been completed.

This considers that only one descriptor is pending at a time. The driver
seems to mixup the DMA descriptors et SG descriptors.